// program/clients/typescript/src/generated/types/policyType.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from '@solana/kit';

export enum PolicyType {
  Refund,
  Settlement,
}

export type PolicyTypeArgs = PolicyType;

export function getPolicyTypeEncoder(): FixedSizeEncoder<PolicyTypeArgs> {
  return getEnumEncoder(PolicyType);
}

export function getPolicyTypeDecoder(): FixedSizeDecoder<PolicyType> {
  return getEnumDecoder(PolicyType);
}

export function getPolicyTypeCodec(): FixedSizeCodec<
  PolicyTypeArgs,
  PolicyType
> {
  return combineCodec(getPolicyTypeEncoder(), getPolicyTypeDecoder());
}